FVCalc test 2

+++ FVCalc.java
@@ -2,7 +2,7 @@
 public class FVCalc {
 	public static void main(String[] args){
 	int currentvalue = Integer.parseInt(args[0]);
-	int rate = Integer.parseInt(args[1]);//assumed to be in %
+	double rate = Double.parseDouble(args[1]);//assumed to be in %
 	int n = Integer.parseInt(args[2]);//years
 	double newrate = rate / 100.0 + 1;
 	double futurervalue = currentvalue * Math.pow( newrate , n );
